+++ docs/court_wizard_bg.wasm.d.ts
@@ -2,14 +2,14 @@
 /* eslint-disable */
 export const memory: WebAssembly.Memory;
 export const main: (a: number, b: number) => number;
-export const __wasm_bindgen_func_elem_135005: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_134883: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_135004: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_133346: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_133336: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_135006: (a: number, b: number, c: number, d: number) => void;
-export const __wasm_bindgen_func_elem_130318: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_130297: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_135051: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_134929: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_133392: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_133382: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_135052: (a: number, b: number, c: number, d: number) => void;
+export const __wasm_bindgen_func_elem_135050: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_130364: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_130343: (a: number, b: number) => void;
 export const __wbindgen_export: (a: number, b: number) => number;
 export const __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
 export const __wbindgen_export3: (a: number) => void;